requirements.txt
click==8.1.8
matplotlib==3.10.0
networkx==3.2.1
numpy==2.2.2
pandas==2.2.3
protobuf==5.29.3
python-dotenv==1.0.1
PyYAML==6.0.2
rdkit==2024.9.5
scikit_learn==1.6.1
setuptools==75.6.0
spacy==3.8.4
torch
torch_geometric
torchmetrics==1.3.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
wandb==0.19.6
xmltodict==0.14.2
